requirements.txt
django-qr-code
asgiref==3.3.1
cairocffi==1.2.0
CairoSVG==2.5.0
cffi==1.14.4
cssselect2==0.4.1
defusedxml==0.6.0
dj-static==0.0.6
Django==3.1.4
django-renderpdf==3.0.0
html5lib==1.1
inflect==5.0.2
lxml==4.6.2
Pillow==8.0.1
pycparser==2.20
Pyphen==0.10.0
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2020.5
reportlab==3.5.57
six==1.15.0
sqlparse==0.4.1
static3==0.7.0
tinycss2==1.1.0
WeasyPrint==52.2
webencodings==0.5.1
cherrypy
daemons
django_wsgiserver
django_celery_results
django_celery_beat
jinja2